==> Scripts/01-CommonVariables.R <==
## Common variables to use in scripts.
## Common file to run from multiple scritps
## PBH Feb 2025

# Load Dimensions and Levels ----------

fuel_colors <- c(
  "Wood and Other Biomass" = "#c5b0d5",
  "Biogenic Municipal Waste" = "#c49c94",
  "Hydropower" = "#17becf",
  "Solar Thermal" = "#9edae5",
  "Solar Photovoltaic" = "#bcbd22",
  "Wind" = "#e377c2",
  "Offshore Wind" = "#7f7f7f",
  "Geothermal" = "#f7b6d1",
  "Pumped Storage/Other" = "#d62728",
  "Distributed Generation" = "#1f77b4",
  "Natural Gas Conventional" = "#CD7F32",
  "Natural Gas Combined Cycle" = "#CC5500",
  "Petroleum" = "#9467bd",
  "Coal" = "#8c564b",
  "Nuclear" = "#2ca02c"
)

# EoF
